+++ hd-timeline_item/templates/template.php
@@ -1,6 +1,6 @@
 <?php
 
-/* Herzog Dupont Copyright (C) 2016–2021 YOOtheme GmbH, 2019–2021 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2016–2021 YOOtheme GmbH, 2019–2021 Thomas Weidlich GNU GPL v3 */
 
 // Display
 foreach (['title', 'meta', 'content', 'link'] as $key) {
@@ -10,33 +10,35 @@ foreach (['title', 'meta', 'content', 'link'] as $key) {
 if (!$element['show_image']) { $props['image'] = $props['icon'] = ''; }
 
 // Resets
-if ($props['icon'] && !$props['image']) { $element['panel_card_image'] = ''; }
+if ($props['icon'] && !$props['image']) { $element['panel_image_no_padding'] = ''; }
 
 // Override default settings
 $element['panel_style'] = $props['panel_style'] ?: $element['panel_style'];
 
-// New logic shortcuts
-$element['has_panel_card_image'] = $props['image'] && $element['panel_card_image'] && $element['image_align'] != 'between';
-$element['has_content_padding'] = $props['image'] && $element['panel_content_padding'] && $element['image_align'] != 'between';
-
 // Image
 $props['image'] = $this->render("{$__dir}/template-image", compact('props'));
 
+// New logic shortcuts
+$element['has_panel_image_no_padding'] = $props['image'] && (!$element['panel_style'] || $element['panel_image_no_padding']) && $element['image_align'] != 'between';
+$element['has_no_padding'] = !$element['panel_style'] && (!$props['image'] || ($props['image'] && $element['image_align'] == 'between'));
+
+// Transition
 if ($element['image_transition']) {
 
     $transition_toggle = $this->el('div', [
         'class' => [
             'uk-inline-clip [uk-transition-toggle {@image_link}]',
-            'uk-border-{image_border}' => !$element['panel_style'] || ($element['panel_style'] && (!$element['panel_card_image'] || $element['image_align'] == 'between')),
+            'uk-border-{image_border}' => !$element['panel_style'] || ($element['panel_style'] && (!$element['panel_image_no_padding'] || $element['image_align'] == 'between')),
             'uk-box-shadow-{image_box_shadow} {@!panel_style}',
             'uk-box-shadow-hover-{image_hover_box_shadow} {@!panel_style}' => $props['link'] && ($element['image_link'] || $element['panel_link']),
-            'uk-margin[-{image_margin}]-top {@!image_margin: remove} {@!image_box_decoration}' => $element['image_align'] == 'between' || ($element['image_align'] == 'bottom' && !($element['panel_style'] && $element['panel_card_image'])),
+            'uk-margin[-{image_margin}]-top {@!image_margin: remove} {@!image_box_decoration}' => $element['image_align'] == 'between' || ($element['image_align'] == 'bottom' && !($element['panel_style'] && $element['panel_image_no_padding'])),
         ],
     ]);
     $props['image'] = $transition_toggle($element, $props['image']);
 
 }
 
+// Decoration
 if ($props['image'] && $element['image_box_decoration']) {
 
     $decoration = $this->el('div', [
@@ -47,7 +49,7 @@ if ($props['image'] && $element['image_box_decoration']) {
             'tm-box-decoration-{image_box_decoration: default|primary|secondary}',
             'tm-box-decoration-inverse {@image_box_decoration_inverse} {@image_box_decoration: default|primary|secondary}',
             'uk-inline {@!image_box_decoration: |shadow}',
-            'uk-margin[-{image_margin}]-top {@!image_margin: remove}' => $element['image_align'] == 'between' || ($element['image_align'] == 'bottom' && !($element['panel_style'] && $element['panel_card_image'])),
+            'uk-margin[-{image_margin}]-top {@!image_margin: remove}' => $element['image_align'] == 'between' || ($element['image_align'] == 'bottom' && !($element['panel_style'] && $element['panel_image_no_padding'])),
         ],
 
     ]);
@@ -58,19 +60,20 @@ if ($props['image'] && $element['image_box_decoration']) {
 // Timeline icon
 $props['timeline_icon'] = $this->render("{$__dir}/template-timeline_icon", compact('props'));
 
-// Panel/Card
+// Panel/Card/Tile
 $el = $this->el($props['link'] && $element['panel_link'] ? 'a' : 'div', [
 
     'class' => [
         'el-item',
-        'uk-margin[-{item_margin}]-top',
         'uk-margin-auto uk-width-{item_maxwidth}',
-        'uk-panel {@!panel_style}',
-        'uk-card uk-{panel_style} [uk-card-{panel_size}]',
-        'uk-card-hover {@!panel_style: |card-hover} {@panel_link}' => $props['link'],
-        'uk-card-body {@panel_style} {@!has_panel_card_image}',
-        'uk-margin-remove-first-child' => (!$element['panel_style'] && !$element['has_content_padding']) || ($element['panel_style'] && !$element['has_panel_card_image']),
-        'uk-flex {@panel_style} {@has_panel_card_image} {@image_align: left|right}', // Let images cover the card height if the cards have different heights
+        'uk-panel [uk-{panel_style: tile-.*}] {@panel_style: |tile-.*}',
+        'uk-card uk-{panel_style: card-.*} [uk-card-{!panel_padding: |default}]',
+        'uk-tile-hover {@panel_style: tile-.*} {@panel_link}' => $props['link'],
+        'uk-card-hover {@!panel_style: |card-hover|tile-.*} {@panel_link}' => $props['link'],
+        'uk-padding[-{!panel_padding: default}] {@panel_style: |tile-.*} {@panel_padding} {@!has_panel_image_no_padding} {@!has_no_padding}',
+        'uk-card-body {@panel_style: card-.*} {@panel_padding} {@!has_panel_image_no_padding} {@!has_no_padding}',
+        'uk-margin-remove-first-child' => !in_array($element['image_align'], ['left', 'right']) || !($element['panel_padding'] && $element['has_panel_image_no_padding']),
+        'uk-flex {@panel_style} {@has_panel_image_no_padding} {@image_align: left|right}', // Let images cover the card/tile height if they have different heights
         'uk-transition-toggle {@image_transition} {@panel_link}' => $props['image'],
     ],
 
@@ -81,12 +84,12 @@ $grid = $this->el('div', [
 
     'class' => [
         'uk-child-width-expand',
-        $element['panel_style'] && $element['has_panel_card_image']
+        $element['panel_style'] && $element['has_panel_image_no_padding']
             ? 'uk-grid-collapse uk-grid-match'
             : ($element['image_grid_column_gap'] == $element['image_grid_row_gap']
                 ? 'uk-grid-{image_grid_column_gap}'
                 : '[uk-grid-column-{image_grid_column_gap}] [uk-grid-row-{image_grid_row_gap}]'),
-        'uk-flex-middle {@image_vertical_align}',
+        'uk-flex-middle {@image_vertical_align}' => !($element['panel_style'] && $element['panel_image_no_padding']),
     ],
 
     'uk-grid' => true,
@@ -105,10 +108,11 @@ $cell_image = $this->el('div', [
 $content = $this->el('div', [
 
     'class' => [
-        'uk-card-body uk-margin-remove-first-child {@panel_style} {@has_panel_card_image}',
-        'uk-padding[-{!panel_content_padding: |default}] uk-margin-remove-first-child {@!panel_style} {@has_content_padding}',
+        'uk-padding[-{!panel_padding: default}] {@panel_style: |tile-.*} {@panel_padding} {@has_panel_image_no_padding}',
+        'uk-card-body {@panel_style: card-.*} {@panel_padding} {@has_panel_image_no_padding}',
+        'uk-margin-remove-first-child {@panel_padding} {@has_panel_image_no_padding}',
         // 1 Column Content Width
-        'uk-container uk-container-{panel_content_width}' => $props['image'] && $element['image_align'] == 'top' && !$element['panel_style'] && !$element['panel_content_padding'] && !$element['item_maxwidth'] && (!$element['align_default'] || $element['align_default'] == '1') && (!$element['align_small'] || $element['align_small'] == '1') && (!$element['align_medium'] || $element['align_medium'] == '1') && (!$element['align_large'] || $element['align_large'] == '1') && (!$element['align_xlarge'] || $element['align_xlarge'] == '1'),
+        'uk-container uk-container-{panel_content_width}' => $props['image'] && $element['image_align'] == 'top' && !$element['panel_style'] && !$element['panel_padding'] && !$element['item_maxwidth'] && (!$element['grid_default'] || $element['grid_default'] == '1') && (!$element['grid_small'] || $element['grid_small'] == '1') && (!$element['grid_medium'] || $element['grid_medium'] == '1') && (!$element['grid_large'] || $element['grid_large'] == '1') && (!$element['grid_xlarge'] || $element['grid_xlarge'] == '1'),
     ],
 
 ]);
@@ -116,7 +120,8 @@ $content = $this->el('div', [
 $cell_content = $this->el('div', [
 
     'class' => [
-        'uk-margin-remove-first-child' => (!$element['panel_style'] && !$element['has_content_padding']) || ($element['panel_style'] && !$element['has_panel_card_image']),
+        'uk-margin-remove-first-child' => !($element['panel_padding'] && $element['has_panel_image_no_padding']),
+        'uk-flex uk-flex-middle {@image_vertical_align}' => $element['panel_style'] && $element['panel_image_no_padding'],
     ],
 
 ]);
@@ -125,16 +130,16 @@ $cell_content = $this->el('div', [
 $link = include "{$__dir}/template-link.php";
 
 // Card media
-if ($element['panel_style'] && $element['has_panel_card_image']) {
+if ($element['panel_style'] && $element['has_panel_image_no_padding']) {
     $props['image'] = $this->el('div', [
-        
+
         'class' => [
-            'uk-card-media-{image_align}',
+            'uk-card-media-{image_align} {@panel_style: card-.*}',
             'uk-cover-container{@image_align: left|right}',
         ],
 
         'uk-toggle' => [
-            'cls: uk-card-media-{image_align} uk-card-media-top; mode: media; media: @{image_grid_breakpoint} {@image_align: left|right}',
+            'cls: uk-card-media-{image_align} uk-card-media-top; mode: media; media: @{image_grid_breakpoint} {@image_align: left|right} {@panel_style: card-.*}',
         ],
 
     ], $props['image'])->render($element);
@@ -144,7 +149,7 @@ if ($element['panel_style'] && $element['has_panel_card_image']) {
 
 <?= $props['timeline_icon'] ?>
 
-<div class="timeline-item-container">
+<div class="hd-timeline-item-container">
 
 <?= $el($element, $attrs) ?>